+++ src/lib/videoUtils.ts
@@ -1,0 +1,187 @@
+import * as VIAM from "@viamrobotics/sdk";
+
+interface RunData {
+  success: boolean;
+  err_string?: string;
+  start: string;
+  end: string;
+  duration_ms: number;
+  runs: any[][];
+  readings?: any;
+}
+
+export const handleVideoStoreCommand = async (
+  videoStoreClient: VIAM.GenericComponentClient,
+  runData?: RunData | null
+) => {
+  console.log("handleVideoStoreCommand called");
+
+  if (!videoStoreClient) {
+    console.log("No videoStoreClient available, returning early");
+    return;
+  }
+
+  try {
+    console.log("Creating storage state command...");
+    const storageStateCommand = VIAM.Struct.fromJson({
+      "command": "get-storage-state"
+    });
+    console.log("Storage state command created:", storageStateCommand);
+
+    console.log("Executing storage state command...");
+    const storageResponse = await videoStoreClient.doCommand(storageStateCommand);
+    console.log("Storage state response:", storageResponse);
+
+    const parseCustomTimeFormat = (timeStr: string): Date => {
+      const isoString = timeStr.replace('_', 'T').replace(/-(\d{2})-(\d{2})Z$/, ':$1:$2Z');
+      return new Date(isoString);
+    };
+
+    const formatToCustomTime = (date: Date): string => {
+      return date.toISOString().replace('T', '_').replace(/:/g, '-').replace(/\.\d{3}Z$/, 'Z');
+    };
+
+    const storageData = storageResponse as any;
+    let fromTime: string;
+    let toTime: string;
+
+    if (storageData && storageData.stored_video && storageData.stored_video.length > 0) {
+      const firstVideo = storageData.stored_video[0];
+
+      if (firstVideo.start && firstVideo.end) {
+        fromTime = firstVideo.start;
+        toTime = firstVideo.end;
+      } else if (firstVideo.from && firstVideo.to) {
+        fromTime = firstVideo.from;
+        toTime = firstVideo.to;
+      } else if (firstVideo.time_range) {
+        fromTime = firstVideo.time_range.start || firstVideo.time_range.from;
+        toTime = firstVideo.time_range.end || firstVideo.time_range.to;
+      } else {
+        throw new Error("No valid time range found in stored_video");
+      }
+
+      console.log("Full time range from stored_video:", fromTime, "to", toTime);
+
+      const startTime = parseCustomTimeFormat(fromTime);
+      const endTime = new Date(startTime.getTime() + 5 * 60 * 1000);
+
+      const maxEndTime = parseCustomTimeFormat(toTime);
+      if (endTime > maxEndTime) {
+        toTime = toTime;
+      } else {
+        toTime = formatToCustomTime(endTime);
+      }
+
+      console.log("Using shortened time range for testing:", fromTime, "to", toTime);
+    } else if (storageData && storageData.ranges && storageData.ranges.length > 0) {
+      const firstRange = storageData.ranges[0];
+      fromTime = firstRange.start || firstRange.from;
+      toTime = firstRange.end || firstRange.to;
+      console.log("Using time range from ranges:", fromTime, "to", toTime);
+    } else {
+      if (runData && runData.start && runData.end) {
+        fromTime = runData.start;
+        toTime = runData.end;
+        console.log("Using time range from runData:", fromTime, "to", toTime);
+      } else {
+        throw new Error("No valid time ranges found in storage response and no runData available");
+      }
+    }
+
+    console.log("Creating fetch command...");
+    const fetchCommand = VIAM.Struct.fromJson({
+      "command": "fetch",
+      "from": fromTime,
+      "to": toTime
+    });
+    console.log("Fetch command created:", fetchCommand);
+
+    console.log("Executing fetch command with 30 second timeout...");
+
+    const fetchPromise = videoStoreClient.doCommand(fetchCommand);
+    const timeoutPromise = new Promise((_, reject) => {
+      setTimeout(() => reject(new Error('Fetch command timed out after 30 seconds')), 30000);
+    });
+
+    const fetchResponse = await Promise.race([fetchPromise, timeoutPromise]);
+    console.log("Fetch command response:", fetchResponse);
+
+    const responseObj = fetchResponse as { video?: string };
+    if (responseObj && responseObj.video) {
+      convertBase64ToMp4(responseObj.video, 'fetched_video.mp4');
+    } else {
+      console.log("No video data in response");
+    }
+
+    console.log("handleVideoStoreCommand completed successfully");
+  } catch (error) {
+    console.error("Error executing video store commands:", error);
+    console.log("Error details:", {
+      message: error instanceof Error ? error.message : 'Unknown error',
+      stack: error instanceof Error ? error.stack : undefined
+    });
+  }
+};
+
+export const convertBase64ToMp4 = (base64Data: string, filename: string) => {
+  try {
+    console.log("Converting base64 to MP4...");
+
+    const base64String = base64Data.includes(',')
+      ? base64Data.split(',')[1]
+      : base64Data;
+
+    const binaryString = atob(base64String);
+    const bytes = new Uint8Array(binaryString.length);
+
+    for (let i = 0; i < binaryString.length; i++) {
+      bytes[i] = binaryString.charCodeAt(i);
+    }
+
+    const blob = new Blob([bytes], { type: 'video/mp4' });
+
+    const url = URL.createObjectURL(blob);
+    const link = document.createElement('a');
+    link.href = url;
+    link.download = filename;
+
+    document.body.appendChild(link);
+    link.click();
+
+    document.body.removeChild(link);
+    URL.revokeObjectURL(url);
+
+    console.log(`MP4 file "${filename}" downloaded successfully`);
+  } catch (error) {
+    console.error("Error converting base64 to MP4:", error);
+  }
+};
+
+export const formatDuration = (durationMs?: number, start?: string, end?: string): string => {
+  let ms = durationMs;
+
+  if (!ms && start && end) {
+    ms = new Date(end).getTime() - new Date(start).getTime();
+  }
+
+  if (!ms) return '0:00';
+
+  const seconds = Math.floor(ms / 1000);
+  const minutes = Math.floor(seconds / 60);
+  const remainingSeconds = seconds % 60;
+  return `${minutes}:${remainingSeconds.toString().padStart(2, '0')}`;
+};
+
+export const formatTimestamp = (timestamp: string): string => {
+  return new Date(timestamp).toLocaleString();
+};
+
+export const formatShortTimestamp = (timestamp: string): string => {
+  return new Date(timestamp).toLocaleTimeString();
+};
+
+export const extractCameraName = (filename: string): string => {
+  const match = filename.match(/video_([^/]+)/);
+  return match ? `${match[1]}` : 'Unknown Camera';
+};
